Fix: update current month usage HTML parsing
The layout for the individual month usage values has changed, which
breaks the parsing. This patch fixes the parsing by removing the
failing intermediate step (using a `> br` selector which doesn't
resolve to anything in the new layout), and simply runs the regex to
extract the usage data on the entire span element.

Signed-off-by: Antoine Busque <antoinebusque@gmail.com>