Fix: update selectors for compatibility with BeautifulSoup 4.7.x
As of Beautiful Soup 4.7.0 [0], the Beautiful Soup CSS selector has
been replaced by that of the Soup Sieve project [1]. This
implementation is stricter than the older Beautiful Soup
implementation, which causes failures when one or more selectors are
skipped, whereas the old implementation would still succeed, provided
the path was unambiguous.

Therefore, update our selectors to work with this new implementation.

[0] https://bazaar.launchpad.net/~leonardr/beautifulsoup/bs4/view/head:/CHANGELOG
[1] https://pypi.org/project/soupsieve/

Signed-off-by: Antoine Busque <antoinebusque@gmail.com>